requirements.txt
flask==2.0.1
werkzeug==2.0.2
tensorflow==2.16.1
numpy>=1.24.0
Pillow>=9.0.0 